requirements.txt
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==4.3.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
ecdsa==0.19.1
fastapi==0.129.0
fastapi-utils==0.8.0
greenlet==3.3.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Mako==1.3.10
MarkupSafe==3.0.3
mypy_extensions==1.1.0
passlib==1.7.4
psutil==5.9.8
pyasn1==0.6.2
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.22
redis==7.1.1
requests==2.32.5
rsa==4.9.1
setuptools==82.0.0
simplejson==3.20.2
six==1.17.0
SQLAlchemy==2.0.46
sqlparse==0.5.5
starlette==0.52.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
wakeonlan==3.1.0
